\name{aveobs}

\alias{aveobs}

\title{
    Estimate an average value
}

\description{
    Given observations located in time, average values for various time
    periods (e.g. year, month, week, day) are generated.
}

\usage{
    tmp <- aveobs(time, values, period="month")
}

\arguments{
    \item{time}{Time identification of the values to integrate}
    \item{values}{The values to be integrated}
    \item{period}{Character representation of the period to integrate over}
}

\value{
}

\seealso{
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
