\name{aveobsdepth}

\alias{aveobsdepth}

\title{
    Estimate an average value
}

\description{
    Given observations located in time and depth, average values for
    various depths are generated.
}

\usage{
    tmp <- aveobsdepth(x, depth)
}

\arguments{
    \item{x}{}
    \item{depth}{}
}

\value{
}

\seealso{
    \link{readhydrography}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
