\name{intobsdepth}

\alias{ }

\title{
    Interpolate observations
}

\description{
    Hydrographic observations are interpolated to standard depth levels.
}

\usage{
    intobsdepth(x)
}

\arguments{
    \item{x}{an object (as returned by \link{readhydrography} to interpolate}
}

\value{
}

\seealso{
    \link{readhydrography}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
