\name{ plotbsrnradobs }

\alias{plotbsrnradobs }

\title{
    Plot relevant BSRN stations observing irradiance
}

\description{
    Plot a map of stations operated as Baseline Surface Radiation Network
    stations that observe components of the surface irradiance. This is
    stations observing sunshine duration, short- or longwave iradiance at
    the surface or other components like direct/diffuse irradiance or
    upward irradiance.
}

\usage{
    plotbsrnradobs()
}

\arguments{
    \item{}{}
}

\value{
}

\seealso{
    \link{plotnorwradobs}, \link{plotsmhiradobs}, \link{map}, \link{mapproj}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
