\name{ plotnorwradobs }

\alias{plotnorwradobs }

\title{
    Plot Norwegian stations observing irradiance
}

\description{
    Plot a map of stations owned and operated by Norwegian institutions
    that observe components of the surface irradiance. This is stations
    observing sunshine duration, short- or longwave iradiance at the
    surface or other components like direct/diffuse irradiance or upward
    irradiance.
}

\usage{
    plotnorwradobs()
}

\arguments{
    \item{}{}
}

\value{
}

\seealso{
    \link{plotsmhiradobs}, \link{plotbsrnradobs}, \link{map}, \link{mapproj}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
