\name{ plotsmhiradobs }

\alias{plotsmhiradobs }

\title{
    Plot SMHI stations observing irradiance
}

\description{
    Plot a map of stations operated by the Swedish Meteorological and
    Hydrological Insitute that observe components of the surface
    irradiance. This is stations observing sunshine duration, short- or
    longwave iradiance at the surface or other components like
    direct/diffuse irradiance or upward irradiance.
}

\usage{
    plotsmhiradobs()
}

\arguments{
    \item{}{}
}

\value{
}

\seealso{
    \link{plotnorwradobs}, \link{plotbasrnradobs}, \link{map}, \link{mapproj}
}

\author{
    Øystein Godøy (\email{o.godoy@met.no})
}

\examples{
}

\keyword{
}
